requirements.txt
wtforms
flask
flask-wtf
MySQL-python
passlib
fastecdsa
Flask-QR
python-dotenv